Sending events from client to server
As servers can emit events, client can do the same thing:
You can create an emit event on client, the name of the event is your decision. Then the second parameter is the data you will send to the server (boolean, object, array, string, etc).
You server must listen for an event 'socket.on' with the same name of your client name, then via callback they can control the data sended to the server.
all is working as expected.